Random.Xor128: Add the state initializer
The current implementation of Random.Xor128 did not have any
mixing algorithm for the initial state.
Using a poor initial state deprives the randomness of a generated sequence
severely, and thus the current implementation was very bad in this way.

This change employs the state initializer derived from splitmix64 and
improves the quality of a generated sequence dramatically.