Vim.BufferManager: Fix open() for range: current
The document said if 'range' options is set to 'current', the manager
only check if a current window is managed or not.
However the implementation never check if a current buffer is managed
and open a new window even a curernt window is managed.

This commit fix this bug to check a current window if range isnot 'all' or 'tabpage'